+++ proc.c
@@ -2,4 +2,4 @@
 #include "proc.h"
 
 Process processes[N_PROC];
-uint8_t stack[N_PROC][STACK_SIZE];
+uintptr_t stack[N_PROC][STACK_SIZE];
